MVP-69 refactor: notion api에 env파일 적용
